refactor(wallet-filter-research): update wallet data handling and taxonomy references
- Revised the SKILL.md to clarify the source of truth for wallet data, emphasizing the use of packages/ecosystem-data for all wallet-related information.
- Updated references in filter-taxonomy.md to point to the new wallet data structure.
- Modified fetch_grid_wallets.mjs to streamline the GraphQL query and adjust the data mapping for wallet URLs, removing unnecessary fields and focusing on essential data.
