+++ packages/ecosystem-data/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./companies";
 export * from "./selectors";
+export * from "./wallets";
 export type * from "./types";
