+++ packages/ecosystem-data/src/wallets/wallet-data.ts
@@ -1,48 +1,47 @@
-// import defaultIcon from "../../../assets/wallets/wallet-placeholder-icon.png";
-import backpackIcon from "../../../assets/wallets/icons/backpack.png";
-import binanceIcon from "../../../assets/wallets/icons/binance.png";
-import bitGetIcon from "../../../assets/wallets/icons/bitget.png";
-import braveIcon from "../../../assets/wallets/icons/brave.png";
-import cakewalletIcon from "../../../assets/wallets/icons/cakewallet.png";
-import cdpIcon from "../../../assets/wallets/icons/cdp.png";
-import circleIcon from "../../../assets/wallets/icons/circle.png";
-import coinbaseIcon from "../../../assets/wallets/icons/coinbase.png";
-import coinwalletIcon from "../../../assets/wallets/icons/coinwallet.png";
-import crossmintIcon from "../../../assets/wallets/icons/crossmint.png";
-import cypherockIcon from "../../../assets/wallets/icons/cypherock.png";
-import decafIcon from "../../../assets/wallets/icons/decaf.png";
-import dynamicIcon from "../../../assets/wallets/icons/dynamic.png";
-import enkryptIcon from "../../../assets/wallets/icons/enkrypt.png";
-import exodusIcon from "../../../assets/wallets/icons/exodus.png";
-import fuseIcon from "../../../assets/wallets/icons/fuse.png";
-import gemIcon from "../../../assets/wallets/icons/gem.png";
-import glowIcon from "../../../assets/wallets/icons/glow.png";
-import guardaIcon from "../../../assets/wallets/icons/guarda.png";
-import heliumIcon from "../../../assets/wallets/icons/helium.png";
-import keystoneIcon from "../../../assets/wallets/icons/keystone.png";
-import ledgerIcon from "../../../assets/wallets/icons/ledger.png";
-import metaMaskIcon from "../../../assets/wallets/icons/metamask.png";
-import nowwalletIcon from "../../../assets/wallets/icons/nowwallet.png";
-import okxIcon from "../../../assets/wallets/icons/okx.png";
-import onekeyIcon from "../../../assets/wallets/icons/onekey.png";
-import paraIcon from "../../../assets/wallets/icons/para.png";
-import phantomIcon from "../../../assets/wallets/icons/phantom.png";
-import privyIcon from "../../../assets/wallets/icons/privy.png";
-import robinhoodIcon from "../../../assets/wallets/icons/robinhood.png";
-import solflareIcon from "../../../assets/wallets/icons/solflare.png";
-import tiplinkIcon from "../../../assets/wallets/icons/tiplink.png";
-import tokenPocketIcon from "../../../assets/wallets/icons/tokenpocket.png";
-import trezorIcon from "../../../assets/wallets/icons/trezor.png";
-import trustWalletCoreIcon from "../../../assets/wallets/icons/trustwalletcore.png";
-import turnkeyIcon from "../../../assets/wallets/icons/turnkey.png";
-import walletConnectIcon from "../../../assets/wallets/icons/walletconnect.png";
-import web3authIcon from "../../../assets/wallets/icons/web3auth.png";
-import atomicwalletIcon from "../../../assets/wallets/icons/atomicwallet.png";
-import jupiterIcon from "../../../assets/wallets/icons/jupiter.png";
-import openfortIcon from "../../../assets/wallets/icons/openfort.png";
-import infinexIcon from "../../../assets/wallets/icons/infinex.png";
-import opensignerIcon from "../../../assets/wallets/icons/opensigner.png";
-import samuiIcon from "../../../assets/wallets/icons/samui.png";
+import backpackIcon from "../../assets/wallets/icons/backpack.webp";
+import binanceIcon from "../../assets/wallets/icons/binance.webp";
+import bitGetIcon from "../../assets/wallets/icons/bitget.webp";
+import braveIcon from "../../assets/wallets/icons/brave.webp";
+import cakewalletIcon from "../../assets/wallets/icons/cakewallet.webp";
+import cdpIcon from "../../assets/wallets/icons/cdp.webp";
+import circleIcon from "../../assets/wallets/icons/circle.webp";
+import coinbaseIcon from "../../assets/wallets/icons/coinbase.webp";
+import coinwalletIcon from "../../assets/wallets/icons/coinwallet.webp";
+import crossmintIcon from "../../assets/wallets/icons/crossmint.webp";
+import cypherockIcon from "../../assets/wallets/icons/cypherock.webp";
+import decafIcon from "../../assets/wallets/icons/decaf.webp";
+import dynamicIcon from "../../assets/wallets/icons/dynamic.webp";
+import enkryptIcon from "../../assets/wallets/icons/enkrypt.webp";
+import exodusIcon from "../../assets/wallets/icons/exodus.webp";
+import fuseIcon from "../../assets/wallets/icons/fuse.webp";
+import gemIcon from "../../assets/wallets/icons/gem.webp";
+import glowIcon from "../../assets/wallets/icons/glow.webp";
+import guardaIcon from "../../assets/wallets/icons/guarda.webp";
+import heliumIcon from "../../assets/wallets/icons/helium.webp";
+import keystoneIcon from "../../assets/wallets/icons/keystone.webp";
+import ledgerIcon from "../../assets/wallets/icons/ledger.webp";
+import metaMaskIcon from "../../assets/wallets/icons/metamask.webp";
+import nowwalletIcon from "../../assets/wallets/icons/nowwallet.webp";
+import okxIcon from "../../assets/wallets/icons/okx.webp";
+import onekeyIcon from "../../assets/wallets/icons/onekey.webp";
+import paraIcon from "../../assets/wallets/icons/para.webp";
+import phantomIcon from "../../assets/wallets/icons/phantom.webp";
+import privyIcon from "../../assets/wallets/icons/privy.webp";
+import robinhoodIcon from "../../assets/wallets/icons/robinhood.webp";
+import solflareIcon from "../../assets/wallets/icons/solflare.webp";
+import tiplinkIcon from "../../assets/wallets/icons/tiplink.webp";
+import tokenPocketIcon from "../../assets/wallets/icons/tokenpocket.webp";
+import trezorIcon from "../../assets/wallets/icons/trezor.webp";
+import trustWalletCoreIcon from "../../assets/wallets/icons/trustwalletcore.webp";
+import turnkeyIcon from "../../assets/wallets/icons/turnkey.webp";
+import walletConnectIcon from "../../assets/wallets/icons/walletconnect.webp";
+import web3authIcon from "../../assets/wallets/icons/web3auth.webp";
+import atomicwalletIcon from "../../assets/wallets/icons/atomicwallet.webp";
+import jupiterIcon from "../../assets/wallets/icons/jupiter.webp";
+import openfortIcon from "../../assets/wallets/icons/openfort.webp";
+import infinexIcon from "../../assets/wallets/icons/infinex.webp";
+import opensignerIcon from "../../assets/wallets/icons/opensigner.webp";
+import samuiIcon from "../../assets/wallets/icons/samui.webp";
 
 export const walletData = [
   {
